Added parameter to events; added chords implementation

+++ src/main/kotlin/kLoops/music/typesUtils.kt
@@ -2,7 +2,7 @@ package kLoops.music
 
 import kotlin.math.round
 
-fun Double.toMidiRange() = round(this * 127)
+fun Double.toMidiRange() = round(this * 127).toInt()
 
 operator fun Int.contains(chances: Int): Boolean =
         Math.random() < (chances.toDouble() / this.toDouble())
